core/pyproject.toml
[project]
name = "kamui-core"
version = "1.0.0"
description = "Backend Kamui: API FastAPI, monitor de pasta, upload YouTube"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "watchdog==3.0.0",
    "python-dotenv==1.0.0",
    "google-api-python-client==2.108.0",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
